backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
constantly==15.1.0
cryptography==41.0.5
dateparser==1.2.0
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pyOpenSSL==23.2.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2023.3.post1
regex==2024.4.16
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzlocal==5.2
ujson==5.9.0
urllib3==2.0.7
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
zope.interface==6.1
